+++ 숙제/3개월차/API 시연회/RimWorld/RimWorld/Pawn.cpp
@@ -568,14 +568,14 @@ void CPawn::Render(HDC hDC)
         CNode* pPrevNode = nullptr;
 
         // 회색 펜 생성 (두께 3)
-        HPEN hGrayPen = CreatePen(PS_SOLID, 3 * CCamera::Get_Instance()->Get_Zoom(), RGB(255, 255, 255));
+        HPEN hGrayPen = CreatePen(PS_SOLID, int(3 * CCamera::Get_Instance()->Get_Zoom()), RGB(255, 255, 255));
         HPEN hOldPen = (HPEN)SelectObject(hDC, hGrayPen);
 
         // 노드 리스트 순회
         for (CNode* pNode : m_NodeList) {
             // 현재 노드의 위치
 
-            POINT tPoint = CCamera::Get_Instance()->WorldToScreen(pNode->Get_Pos().iX, pNode->Get_Pos().iY);
+            POINT tPoint = CCamera::Get_Instance()->WorldToScreen((float)(pNode->Get_Pos().iX), (float)(pNode->Get_Pos().iY));
 
             int iCurrentX = int(tPoint.x + iScrollX);
             int iCurrentY = int(tPoint.y + iScrollY);
@@ -583,7 +583,7 @@ void CPawn::Render(HDC hDC)
             // 이전 노드가 있다면 선을 그림
             if (pPrevNode != nullptr) 
             {
-                POINT tPoint = CCamera::Get_Instance()->WorldToScreen(pPrevNode->Get_Pos().iX, pPrevNode->Get_Pos().iY);
+                POINT tPoint = CCamera::Get_Instance()->WorldToScreen((float)(pPrevNode->Get_Pos().iX), (float)(pPrevNode->Get_Pos().iY));
 
                 int iPrevX = int(tPoint.x + iScrollX);
                 int iPrevY = int(tPoint.y + iScrollY);
