+++ 숙제/3개월차/API 시연회/RimWorld/RimWorld/RangedWeapon.cpp
@@ -112,7 +112,7 @@ void CRangedWeapon::Render(HDC hDC)
     int		iScrollX = (int)CScrollMgr::Get_Instance()->Get_ScrollX();
     int		iScrollY = (int)CScrollMgr::Get_Instance()->Get_ScrollY();
 
-    POINT tPoint = CCamera::Get_Instance()->WorldToScreen(m_tRect.left, m_tRect.top);
+    POINT tPoint = CCamera::Get_Instance()->WorldToScreen((float)m_tRect.left, (float)m_tRect.top);
     float fZoom = CCamera::Get_Instance()->Get_Zoom();
